+++ src/lib/openai/completion.rs
@@ -15,7 +15,7 @@ use tokio::spawn;
 use tracing::{debug, error, info, warn};
 
 use crate::{
-    client::AppendHeaders,
+    client::{SUPER_CLIENT, SetupRequest},
     error::{ClewdrError, check_res_err},
     messages::ClientRequestBody,
     openai::stream::{ClewdrTransformer, NonStreamEventData},
@@ -141,7 +141,7 @@ impl AppState {
         print_out_json(&p, "0.req.json");
         let stream = p.stream;
         let proxy = self.config.rquest_proxy.clone();
-        let Some(ref org_uuid) = self.org_uuid else {
+        let Some(org_uuid) = self.org_uuid.clone() else {
             return Ok(Json(json!(
                 {
                     "error": {
@@ -174,15 +174,14 @@ impl AppState {
             body["model"] = p.model.clone().into();
         }
 
-        let api_res = self
-            .client
+        let api_res = SUPER_CLIENT
             .post(endpoint)
             .json(&body)
-            .append_headers("", proxy.clone())
+            .setup_request("", self.header_cookie(), proxy.clone())
             .send()
             .await?;
         debug!("New conversation created: {}", new_uuid);
-
+        self.update_cookie_from_res(&api_res);
         check_res_err(api_res).await?;
 
         // generate the request body
@@ -217,15 +216,14 @@ impl AppState {
             new_uuid
         );
 
-        let api_res = self
-            .client
+        let api_res = SUPER_CLIENT
             .post(endpoint)
             .json(&body)
-            .append_headers("", proxy)
+            .setup_request("", self.header_cookie(), proxy)
             .header_append(ACCEPT, "text/event-stream")
             .send()
             .await?;
-
+        self.update_cookie_from_res(&api_res);
         let api_res = check_res_err(api_res).await?;
 
         if !stream {